cPickle?
Hey guys,

I'll start by saying that I haven't actually tested this yet. But I was wondering... is there any reason why we are not using the C implementation of pickle? It seems to be much faster! And it has happened to me that while the python pickle module died with some data, cPickle kept on going, like the duracell bunny